% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new-repo.R
\name{new_repo_blfs}
\alias{new_repo_blfs}
\title{Start using Box LFS on a new project}
\usage{
new_repo_blfs(dir = NULL, size = 10, box_dir = NULL, boxdrive = TRUE)
}
\arguments{
\item{dir}{the file path to the file directory}

\item{size}{the minimum file size in megabytes to track}

\item{box_dir}{the file path to the project within Box}

\item{boxdrive}{logical, should Box drive be used?}
}
\value{
Creates the following files in \code{dir}:
\itemize{
\item box-lfs: to store the .boxtracker files
\item upload: put copies of tracked files for easy uploading to Box
\item *.boxtracker: the tracker files for each large file
}

Also adds the tracked files and the /box-lfs/upload file to .gitignore.
}
\description{
Use if you have an existing directory that you want to start tracking with git and put on GitHub. It will set up the
file structure, identify files that should be tracked, and prompt user to upload those files to Box.
}
\details{
Box LFS works similar to git lfs where large files are tracked using a tracking file (.boxtracker) which keeps track of the file location and
it's version. Any files larger than the specified size (default is 10 MB) will be added to .gitignore and the user will be prompted to upload those files
to Box and supply the Box link to those files.
}
\examples{
tmp <- withr::local_tempdir()
new_repo_blfs(tmp)
}
