% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/untrack-files.R
\name{rm_tracking}
\alias{rm_tracking}
\title{Stop tracking a file with Box LFS}
\usage{
rm_tracking(dir, file, git = TRUE, box = FALSE)
}
\arguments{
\item{dir}{the file path to the file directory}

\item{file}{the relative file path to the file to stop tracking}

\item{git}{logical, should the file be tracked by git again?}

\item{box}{logical, should the file be removed from Box?}
}
\value{
a message indicating that the file has been successfully untracked
}
\description{
Removes the boxtracker file, adds to .blfsignore, and optionally removes it from
.gitignore so that git will start tracking it again.
}
\examples{
#create temp dir to modify files cleanly
  tmp <- blfs:::create_test_repo(box_lfs=TRUE, examples = TRUE,
  source_dir = system.file("extdata", package = "blfs"))

  rm_tracking(tmp, "example-files/large-file1.txt")
  rm_tracking(tmp, "example-files/example-shp.shp", git=FALSE)

  #remove temp dirs
  unlink(tmp, recursive = TRUE)
}
